Removing symlink for non-osx platforms

+++ do.py
@@ -868,12 +868,13 @@ def _build_rw_top_trumps_lib():
             # Create a symlink to the library to be used at run-time
             library_src = os.path.abspath(os.path.join('code-experiments', 'rw-problems',
                                                        'top_trumps', rw_library))
-            library_des = '/usr/local/lib/' + rw_library
-            if os.path.lexists(library_des):
-                os.remove(library_des)
-            os.symlink(library_src, library_des)
             # Copy the library so that the socket server finds it
             copy_file(library_src, os.path.join('code-experiments', 'rw-problems', rw_library))
+            if 'darwin' in sys.platform:
+                library_des = '/usr/local/lib/' + rw_library
+                if os.path.lexists(library_des):
+                    os.remove(library_des)
+                os.symlink(library_src, library_des)
     except PermissionError as e:
         print('Encountered a permission error, the rw-top-trumps library is probably used by the'
               'server. Stop the socket server and try again. \nError: {}'.format(e))
